// packages/verino/src/core/types.ts
export type InputType = 'numeric' | 'alphabet' | 'alphanumeric' | 'any'